Реглы добавил поле файлов

+++ routes/param.route.js
@@ -4,7 +4,72 @@ const auth = require('../middleware/middleware.auth')
 
 const Param = require('../models/Param')
 
-// /param
+async function changeOrder(newOrder, oldOrder, paramPage){
+
+  if (newOrder < 1){
+    return {message: `Порядок отображения заголовка не может быть меньше 0!`}
+  }
+
+  const orderLength = await Param.find({page: paramPage}).countDocuments()
+
+  if (newOrder > orderLength) {
+    return {message: `Порядок отображения заголовка для данной страницы должен быть меньше или равен ${orderLength}`}
+  }
+
+  if(newOrder && newOrder !== oldOrder){
+    if (newOrder > oldOrder){
+
+      await Param.updateMany({
+        page: paramPage,
+        order: {$lte: newOrder, $gt: oldOrder}
+      },{
+        $inc: { order: -1 }
+      })
+
+    } else if (newOrder < oldOrder){
+
+      await Param.updateMany({
+        page: paramPage, 
+        order: {$lt: oldOrder, $gte: newOrder}
+      }, {
+        $inc:{ order: 1 }
+      })
+
+    }
+  }
+
+  return false
+
+}
+
+async function updateOrdersOnPage(page, order, method){
+
+  if (order < 1){
+    return {message: `Порядок отображения заголовка не может быть меньше 0!`}
+  }
+
+  const orderLength = await Param.find({page}).countDocuments()
+
+  if (order > orderLength) {
+    return {message: `Порядок отображения заголовка для данной страницы должен быть меньше или равен ${orderLength}`}
+  }
+
+  await Param.updateMany({
+    page, 
+    order: {$gte: order}
+  }, {
+    $inc:{ order: method === 'post' ? 1 : -1 }
+  })
+
+  return false
+
+}
+
+/**
+ * @url /api/param/
+ * @description Получить все заголовки
+ * @param page
+ */
 router.get('/', async (req, res) => {
 
   const page = req.query.page
@@ -20,6 +85,10 @@ router.get('/', async (req, res) => {
   }
 })
 
+/**
+ * @description Получить все активные заголовки
+ * @param page
+ */
 router.get('/active', async(req,res)=>{
   
   const page = req.query.page
@@ -42,7 +111,10 @@ router.get('/active', async(req,res)=>{
   }
 })
 
-// /param/:id
+/**
+ * @description Получить заголовок по id
+ * @param id
+ */
 router.get('/:id', async (req, res) => {
 
   const id = req.params.id
@@ -61,7 +133,9 @@ router.get('/:id', async (req, res) => {
   }
 })
 
-// param/
+/**
+ * @description Создать заголовок
+ */
 router.post('/', auth, async (req, res) => {
 
   const {
@@ -83,21 +157,26 @@ router.post('/', auth, async (req, res) => {
 
     const param = new Param({ title, page, text, isActive, order, img })
 
-    // console.log(exists);
+    const updated = await updateOrdersOnPage(page, order, 'post')
+    if (updated.message){
+      return res.status(400).json({message: updated.message})
+    }
 
     await param.save()
-        .then(param => res.json({ message: "Заголовок успешно добавлен", param }))
-        .catch(err => res.status(400).json({ message: err.message }))
+      .then(param => res.json({ message: "Заголовок успешно добавлен", param }))
+      .catch(err => res.status(400).json({ message: err.message }))
 
   } catch (error) {
-    res.status(500).json({ message: error.message })
+    return res.status(500).json({ message: error.message })
   }
 })
 
-// param/:id
+/**
+ * @description удалить заголовок по id
+ */
 router.delete('/:id', auth, async (req, res) => {
 
-  const id = req.params.id
+  const {id} = req.params
 
   try {
     const param = await Param.findById(id)
@@ -105,6 +184,12 @@ router.delete('/:id', auth, async (req, res) => {
     if (!param) {
       return res.status(404).json({ message: "Заголовок не найден" })
     }
+
+    const updated = await updateOrdersOnPage(param.page, param.order, 'delete')
+    if (updated.message){
+      return res.status(400).json({message: updated.message})
+    }
+
     await param.delete()
       .then(() => res.json({ message: "Заголовок удален" }))
 
@@ -113,126 +198,87 @@ router.delete('/:id', auth, async (req, res) => {
   }
 })
 
-router.put('/:id', async(req,res)=>{ // only for changing order
+/**
+ * @url /api/param/:id
+ * @description Изменить порядок или отображение заголовка
+ * @param isActive
+ * @param order
+ */
+router.put('/:id', async (req,res) => { 
+
   const id = req.params.id
   const {order, isActive} = req.body
+
   try{
-    console.log(order, isActive);
     
-    if(!order && (isActive===undefined)){
-      return res.status(422).json({message: "Не задан порядок отображения или статус активности. Попробуйте еще раз"})
+    if(!order && !isActive){
+      return res.status(400).json({message: "Не задан порядок отображения или статус активности. Попробуйте еще раз"})
     } 
 
     const param = await Param.findById(id)
-
-    const order_length = await Param.find({page: param.page}).countDocuments()
-
-    if (order > order_length) {
-      return res.status(422).json({message: `Порядок отображения должен быть меньше или равен ${order_length} для данной страницы`})
-    }
-
+    
     if (!param) {
       return res.status(404).json({ message: "Заголовок не найден" })
     }
 
-    // const result = await Param.find({page: param.page, order:{$gte: param.order, $lt:order}})
-    // const result_2 = await Param.find({page: param.page, order:{$gt: order, $lte:param.order}})
+    const updated = await changeOrder(order, param.order, param.page)
 
-    // console.log(result, result_2);
-    
-
-    if(order){
-      if (order !== param.order){
-        if (order > param.order){
-          await Param.updateMany({
-            page: param.page,
-            order: {$lte: order, $gte: param.order}
-          },{
-            $inc: { order: -1}
-          })
-        } else if (order < param.order){
-          await Param.updateMany({
-            page: param.page, 
-            order: {$lte: param.order, $gte: order}
-          }, {
-            $inc:{ order: 1}
-          })
-        }
-      }
-  
-      param.order = order
+    if(updated.message){
+      return res.status(400).json({message: updated.message})
     }
 
-    if((isActive !== param.isActive) && isActive!==undefined){
+    param.order = order
+    
+    if((isActive !== param.isActive) && isActive){
       param.isActive = isActive
     }
 
     await param.save()
-      .then(param=>res.json({message: `Порядок заголовка ${param.title} на странице ${param.page}: ${param.order}, статус активности: ${param.isActive? `виден` : `скрыт`}`}))
-
-    // things
+      .then(param=>res.json({message: `Порядок заголовка "${param.title}" на странице "${param.page}": ${param.order}, статус активности: ${param.isActive? `виден` : `скрыт`}`}))
 
   } catch (error) {
     res.status(500).json({ message: error.message })
   }
 })
 
+/**
+ * @description Изменить заголовок
+ * @param id
+ */
 router.patch('/:id', async (req, res) => {
 
-  const id = req.params.id
+  const {id} = req.params
   const { page, text, title, isActive, order, img } = req.body
 
   try {
+
+    if (!page){return res.status(400).json({ message: "Поле страница является обязательным" })}
+    if (!title){return res.status(400).json({ message: "Поле заголовок является обязательным" })}
+    if (!text){return res.status(400).json({ message: "Поле текст является обязательным" })}
+
     const param = await Param.findById(id)
 
     if (!param) {
       return res.status(404).json({ message: "Заголовок не найден" })
     }
 
-    if (title || page){
-      // check already exist by page (new\old page) and title (new\old title)
-      const exists = await Param.findOne({
-        title: title !== param.title? title : '',
-        page: page!==param.page? page : ''
-      })
-      if (exists){
-        return res.status(404).json({ message: "Заголовок с таким названием на введенной странице уже существует" })
-      }
+    const exists = await Param.findOne({title, page})
+
+    if (exists && (exists !== param)){
+      return res.status(400).json({ message: "Заголовок с таким названием на данной странице уже существует" })
     }
 
-    if (page && page !== param.page) { param.page = page }
-    if (title && title !== param.title) { param.title = title}
-    if (text && text !== param.text) { param.text = text }
-    if (isActive && isActive!== param.isActive) { param.isActive = isActive }
+    param.page = page
+    param.title = title
+    param.text = text
+    param.isActive = isActive
+    param.img = img ? img : null
 
-    if (img) { 
-      param.img = img 
-    } else {
-      // delete img prop 
-      param.img = null 
+    const updated = await changeOrder(order, param.order, param.page)
+    if(updated.message){
+      return res.status(400).json({message: updated.message})
     }
-
-    if (order && order !== param.order) {       
-
-      if (order > param.order){
-        await Param.updateMany({
-          page: param.page,
-          order: {$lte: order, $gte: param.order}
-        },{
-          $inc: { order: -1}
-        })
-      } else if (order < param.order){
-        await Param.updateMany({
-          page: param.page, 
-          order: {$lte: param.order, $gte: order}
-        }, {
-          $inc:{ order: 1}
-        })
-      }
-  
-      param.order = order  
-      // and other thing to change
-    }
+    param.order = order
 
     await param.save()
       .then(param => res.json({ message: "Заголовок обновлен", param }))
